+++ packages/components/modules/__shared__/native/index.ts
@@ -1,0 +1,2 @@
+export { default as EmptyView } from './EmptyView'
+export type * from './EmptyView/types'