+++ packages/components/modules/messages/native/RoomsList/types.ts
@@ -1,0 +1,7 @@
+import { ChatRoomsQuery$data } from '../../../../__generated__/ChatRoomsQuery.graphql'
+
+export interface RoomsListProps {
+  targetRef: ChatRoomsQuery$data
+  searchParam: string
+  selectedTab: string
+}