+++ packages/components/modules/messages/native/index.ts
@@ -1,3 +1,6 @@
 // export native messages components
 export * from './graphql/subscriptions/useMessagesListSubscription'
 export * from './graphql/subscriptions/useRoomListSubscription'
+export { default as ChatRooms } from './ChatRooms'
+export { default as RoomsList } from './RoomsList'
+export { default as SearchNotFoundState } from './SearchNotFoundState'